Removed mail middleware and moved it to utils
